+++ cmd/cli_test.go
@@ -2,15 +2,11 @@ package main
 
 import (
 	"bytes"
-	"github.com/rendon/testcli"
 	"github.com/statping/statping/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"io/ioutil"
-	"os"
-	"os/exec"
 	"testing"
-	"time"
 )
 
 var (
@@ -18,102 +14,31 @@ var (
 )
 
 func init() {
-	dir = utils.Directory
-	//core.SampleHits = 480
+	utils.InitCLI()
 }
 
-func TestStartServerCommand(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(60*time.Second), got)
-	os.Unsetenv("DB_CONN")
-	gg, _ := <-got
-	assert.Contains(t, gg, "DB_CONN environment variable was found")
-	assert.Contains(t, gg, "Core database does not exist, creating now!")
-	assert.Contains(t, gg, "Starting monitoring process for 5 Services")
+func TestStatpingDirectory(t *testing.T) {
+	dir := utils.Directory
+	require.NotContains(t, dir, "/cmd")
+	require.NotEmpty(t, dir)
+
+	dir = utils.Params.GetString("STATPING_DIR")
+	require.NotContains(t, dir, "/cmd")
+	require.NotEmpty(t, dir)
 }
 
-func TestVersionCommand(t *testing.T) {
-	c := testcli.Command("statping", "version")
-	c.Run()
-	assert.True(t, c.StdoutContains(VERSION))
-}
-
-func TestHelpCommand(t *testing.T) {
-	c := testcli.Command("statping", "help")
-	c.Run()
-	t.Log(c.Stdout())
-	assert.True(t, c.StdoutContains("statping help               - Shows the user basic information about Statping"))
-}
-
-func TestStaticCommand(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "static")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(10*time.Second), got)
-	gg, _ := <-got
-	t.Log(gg)
-	assert.Contains(t, gg, "Exporting Static 'index.html' page...")
-	assert.Contains(t, gg, "Exported Statping index page: 'index.html'")
-	assert.FileExists(t, dir+"/index.html")
-}
-
-func TestExportCommand(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "export")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(10*time.Second), got)
-	gg, _ := <-got
-	t.Log(gg)
-	assert.FileExists(t, dir+"/statping-export.json")
-}
-
-func TestUpdateCommand(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "version")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(15*time.Second), got)
-	gg, _ := <-got
-	t.Log(gg)
-	assert.Contains(t, gg, VERSION)
-}
-
-func TestAssetsCommand(t *testing.T) {
-	t.SkipNow()
-	c := testcli.Command("statping", "assets")
-	c.Run()
-	t.Log(c.Stdout())
-	t.Log("Directory for Assets: ", dir)
-	time.Sleep(1 * time.Second)
-	err := utils.DeleteDirectory(dir + "/assets")
+func TestEnvCLI(t *testing.T) {
+	cmd := rootCmd
+	b := bytes.NewBufferString("")
+	cmd.SetOut(b)
+	cmd.SetArgs([]string{"env"})
+	err := cmd.Execute()
 	require.Nil(t, err)
-	assert.FileExists(t, dir+"/assets/robots.txt")
-	assert.FileExists(t, dir+"/assets/scss/base.scss")
-	assert.FileExists(t, dir+"/assets/scss/main.scss")
-	assert.FileExists(t, dir+"/assets/scss/variables.scss")
-	assert.FileExists(t, dir+"/assets/css/main.css")
-	assert.FileExists(t, dir+"/assets/css/vendor.css")
-	assert.FileExists(t, dir+"/assets/css/style.css")
-	err = utils.DeleteDirectory(dir + "/assets")
+	out, err := ioutil.ReadAll(b)
 	require.Nil(t, err)
-}
-
-func TestRunCommand(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "run")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(15*time.Second), got)
-	gg, _ := <-got
-	t.Log(gg)
-	assert.Contains(t, gg, "Running 1 time and saving to database...")
-	assert.Contains(t, gg, "Check is complete.")
-}
-
-func TestEnvironmentVarsCommand(t *testing.T) {
-	c := testcli.Command("statping", "env")
-	c.Run()
-	assert.True(t, c.StdoutContains("Statping Environment Variable"))
+	assert.Contains(t, string(out), VERSION)
+	assert.Contains(t, utils.Directory, string(out))
+	assert.Contains(t, "SAMPLE_DATA=true", string(out))
 }
 
 func TestVersionCLI(t *testing.T) {
@@ -121,10 +46,11 @@ func TestVersionCLI(t *testing.T) {
 	b := bytes.NewBufferString("")
 	cmd.SetOut(b)
 	cmd.SetArgs([]string{"version"})
-	cmd.Execute()
+	err := cmd.Execute()
+	require.Nil(t, err)
 	out, err := ioutil.ReadAll(b)
-	assert.Nil(t, err)
-	assert.Contains(t, string(out), VERSION)
+	require.Nil(t, err)
+	assert.Contains(t, VERSION, string(out))
 }
 
 func TestAssetsCLI(t *testing.T) {
@@ -132,51 +58,51 @@ func TestAssetsCLI(t *testing.T) {
 	b := bytes.NewBufferString("")
 	cmd.SetOut(b)
 	cmd.SetArgs([]string{"assets"})
-	cmd.Execute()
+	err := cmd.Execute()
+	require.Nil(t, err)
 	out, err := ioutil.ReadAll(b)
 	assert.Nil(t, err)
 	assert.Contains(t, string(out), VERSION)
-	assert.FileExists(t, dir+"/assets/css/main.css")
-	assert.FileExists(t, dir+"/assets/css/style.css")
-	assert.FileExists(t, dir+"/assets/css/vendor.css")
-	assert.FileExists(t, dir+"/assets/scss/base.scss")
-	assert.FileExists(t, dir+"/assets/scss/mobile.scss")
-	assert.FileExists(t, dir+"/assets/scss/variables.scss")
+	assert.FileExists(t, utils.Directory+"/assets/css/main.css")
+	assert.FileExists(t, utils.Directory+"/assets/css/style.css")
+	assert.FileExists(t, utils.Directory+"/assets/css/vendor.css")
+	assert.FileExists(t, utils.Directory+"/assets/scss/base.scss")
+	assert.FileExists(t, utils.Directory+"/assets/scss/mobile.scss")
+	assert.FileExists(t, utils.Directory+"/assets/scss/variables.scss")
 }
 
-func TestSassCLI(t *testing.T) {
-	c := testcli.Command("statping", "assets")
-	c.Run()
-	t.Log(c.Stdout())
-	assert.FileExists(t, dir+"/assets/css/main.css")
-	assert.FileExists(t, dir+"/assets/css/style.css")
-	assert.FileExists(t, dir+"/assets/css/vendor.css")
+func TestHelpCLI(t *testing.T) {
+	cmd := rootCmd
+	b := bytes.NewBufferString("")
+	cmd.SetOut(b)
+	cmd.SetArgs([]string{"help"})
+	err := cmd.Execute()
+	require.Nil(t, err)
+	out, err := ioutil.ReadAll(b)
+	require.Nil(t, err)
+	assert.Contains(t, string(out), VERSION)
 }
 
-func TestUpdateCLI(t *testing.T) {
-	t.SkipNow()
-	cmd := helperCommand(nil, "update")
-	var got = make(chan string)
-	commandAndSleep(cmd, time.Duration(15*time.Second), got)
-	gg, _ := <-got
-	t.Log(gg)
-	assert.Contains(t, gg, "version")
-}
+func TestResetCLI(t *testing.T) {
+	err := utils.SaveFile(utils.Directory+"/statping.db", []byte("test data"))
+	require.Nil(t, err)
 
-func commandAndSleep(cmd *exec.Cmd, duration time.Duration, out chan<- string) {
-	go func(out chan<- string) {
-		runCommand(cmd, out)
-	}(out)
-	time.Sleep(duration)
-	cmd.Process.Kill()
-}
+	cmd := rootCmd
+	b := bytes.NewBufferString("")
+	cmd.SetOut(b)
+	cmd.SetArgs([]string{"reset"})
+	err = cmd.Execute()
+	require.Nil(t, err)
+	out, err := ioutil.ReadAll(b)
+	require.Nil(t, err)
+	assert.Contains(t, string(out), VERSION)
 
-func helperCommand(envs []string, s ...string) *exec.Cmd {
-	cmd := exec.Command("statping", s...)
-	return cmd
-}
+	assert.NoDirExists(t, utils.Directory+"/assets")
+	assert.NoDirExists(t, utils.Directory+"/logs")
+	assert.NoFileExists(t, utils.Directory+"/config.yml")
+	assert.NoFileExists(t, utils.Directory+"/statping.db")
+	assert.FileExists(t, utils.Directory+"/statping.db.backup")
 
-func runCommand(c *exec.Cmd, out chan<- string) {
-	bout, _ := c.CombinedOutput()
-	out <- string(bout)
+	err = utils.DeleteFile(utils.Directory + "/statping.db.backup")
+	require.Nil(t, err)
 }