+++ handlers/groups.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"github.com/gorilla/mux"
-	"github.com/pkg/errors"
+	"github.com/statping/statping/types/errors"
 	"github.com/statping/statping/types/groups"
 	"github.com/statping/statping/utils"
 	"net/http"
@@ -10,13 +10,21 @@ import (
 
 func findGroup(r *http.Request) (*groups.Group, error) {
 	vars := mux.Vars(r)
+	if utils.NotNumber(vars["id"]) {
+		return nil, errors.NotNumber
+	}
 	id := utils.ToInt(vars["id"])
 	if id == 0 {
-		return nil, errors.New("missing group id")
+		return nil, errors.IDMissing
 	}
 	g, err := groups.Find(id)
 	if err != nil {
 		return nil, err
+	}
+	if !g.Public.Bool {
+		if !IsReadAuthenticated(r) {
+			return nil, errors.NotAuthenticated
+		}
 	}
 	return g, nil
 }
@@ -31,7 +39,7 @@ func apiAllGroupHandler(r *http.Request) interface{} {
 func apiGroupHandler(w http.ResponseWriter, r *http.Request) {
 	group, err := findGroup(r)
 	if err != nil {
-		sendErrorJson(errors.Wrap(err, "group not found"), w, r, http.StatusNotFound)
+		sendErrorJson(err, w, r)
 		return
 	}
 	returnJson(group, w, r)
@@ -41,8 +49,7 @@ func apiGroupHandler(w http.ResponseWriter, r *http.Request) {
 func apiGroupUpdateHandler(w http.ResponseWriter, r *http.Request) {
 	group, err := findGroup(r)
 	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		sendErrorJson(errors.Wrap(err, "group not found"), w, r)
+		sendErrorJson(err, w, r)
 		return
 	}
 
@@ -79,7 +86,7 @@ func apiCreateGroupHandler(w http.ResponseWriter, r *http.Request) {
 func apiGroupDeleteHandler(w http.ResponseWriter, r *http.Request) {
 	group, err := findGroup(r)
 	if err != nil {
-		sendErrorJson(errors.Wrap(err, "group not found"), w, r)
+		sendErrorJson(err, w, r)
 		return
 	}