+++ types/messages/database.go
@@ -1,6 +1,9 @@
 package messages
 
-import "github.com/statping/statping/database"
+import (
+	"github.com/statping/statping/database"
+	"github.com/statping/statping/types/errors"
+)
 
 var db database.Database
 
@@ -11,6 +14,9 @@ func SetDB(database database.Database) {
 func Find(id int64) (*Message, error) {
 	var message Message
 	q := db.Where("id = ?", id).Find(&message)
+	if q.Error() != nil {
+		return nil, errors.Missing(message, id)
+	}
 	return &message, q.Error()
 }