+++ utils/utils_test.go
@@ -201,6 +201,5 @@ func TestConfigLoad(t *testing.T) {
 	assert.Equal(t, "sqlite", s("DB_CONN"))
 	assert.Equal(t, Directory, s("STATPING_DIR"))
 	assert.True(t, b("SAMPLE_DATA"))
-	assert.False(t, b("DISABLE_LOGS"))
 	assert.False(t, b("ALLOW_REPORTS"))
 }